backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
agno
together
opencv-python
mediapipe
deepface
librosa
moviepy
faster-whisper
websockets
redis
sse-starlette
sqlalchemy
celery
motor
pymongo
passlib[bcrypt]
python-multipart
google-generativeai
